% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{class}
\name{ddm-class}
\alias{ddm-class}
\alias{setDDM}
\title{An S4 Class of the ddm object}
\usage{
setDDM(model, population_distribution = NULL)
}
\arguments{
\item{model}{A model object containing the model specification}

\item{population_distribution}{Optional population distribution for parameter
simulations (default NULL).}
}
\value{
An object of class lba containing:
\itemize{
\item The model specification
\item Population distribution (if provided)
}
}
\description{
The \code{ddm} class encapsulates a complete decision diffusion model specification,
using model objects from the \code{ggdmcModel} package.
}
\details{
The setDDM function initialises this model by creating an S4 object
with all necessary components.
}
\section{Slots}{

\describe{
\item{\code{model}}{A model object from \code{ggdmcModel} containing the model specification}

\item{\code{population_distribution}}{The population distribution for parameter simulations
(can be NULL).}
}}

