% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddModel.R
\docType{package}
\name{ddModel}
\alias{ddModel-package}
\alias{ddModel}
\title{The Decision Diffusion Model}
\description{
\pkg{ddModel} Density, distribution and random generation functions for the
Decision Diffusion model (DDM). Supports model specification,
parameter estimation, and likelihood computation.
}
\author{
Yi-Shin Lin <yishinlin001@gmail.com> \cr
}
\keyword{internal}
\keyword{package}
