% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddModel.R
\docType{package}
\name{ddModel}
\alias{ddModel-package}
\alias{ddModel}
\title{The Decision Diffusion Model}
\description{
\pkg{ddModel} provides functions for computing the density, distribution,
and random generation of the Decision Diffusion model (DDM), a widely used
cognitive model for analysing choice and response time data. The package
allows model specification, including the ability to fix, constrain, or vary
parameters across experimental conditions. While it does not include a
built-in optimiser, it supports likelihood evaluation and can be integrated
with external tools for parameter estimation. Functions for simulating
synthetic datasets aer also provided. This package is intended for
researchers modelling speeded decision-making in behavioural and cognitive
experiments. For more information, see
Voss, Rothermund, and Voss (2004) <doi:10.3758/BF03196893>,
Voss and Voss (2007) <doi:10.3758/BF03192967>, and
Ratcliff and McKoon (2008) <doi:10.1162/neco.2008.12-06-420>.
}
\author{
Yi-Shin Lin <yishinlin001@gmail.com> \cr
}
\keyword{internal}
\keyword{package}
